#80 izmenjen nacin zastite ruta

+++ WattApp/server/Server/Controllers/UsersController.cs
@@ -48,13 +48,15 @@ namespace Server.Controllers
 
 
         [HttpGet]
-        [Route("page/{page:int}")]
-        [Authorize(Roles = Roles.AdminPermission)]
-        public async Task<IActionResult> GetPage([FromRoute]int page)
+        [Route("page")]
+        [Authorize(Roles = Roles.AdminOperaterPermission)]
+        public async Task<IActionResult> GetPage([FromQuery]int pageNumber, [FromQuery] int pageSize=20)
         {
             try
             {
-                return Ok(await userService.GetPageOfUsers(page, 20, (user) => true));
+                if (User.IsInRole(Roles.Operater))
+                    return Ok(await userService.GetPageOfUsers(pageNumber, pageSize, (user) => user.RoleId==Roles.ProsumerId));
+                return Ok(await userService.GetPageOfUsers(pageNumber, pageSize, (user) => true));
             }
             catch(HttpRequestException ex)
             {
@@ -76,7 +78,7 @@ namespace Server.Controllers
 
         [HttpGet]
         [Route("{id:long}")]
-        [Authorize(Roles = Roles.AdminPermission)]
+        [Authorize(Roles = Roles.AdminOperaterPermission)]
         public async Task<IActionResult> GetUserById([FromRoute]long id)
         {
             var user = await userService.GetUserById(id);
@@ -84,6 +86,10 @@ namespace Server.Controllers
             {
                 return NotFound(new { message="User with id "+id.ToString()+" doesn\'t exist" });
             }
+            if (user.RoleId == Roles.SuperadminId)
+                return Forbid();
+            if (User.IsInRole(Roles.Operater) && user.RoleId != Roles.ProsumerId)
+                return Forbid();
             return Ok(new UserDetailsDTO(user));
         }
 
@@ -120,13 +126,13 @@ namespace Server.Controllers
 
         [HttpGet]
         [Route("roles")]
-        [Authorize(Roles = Roles.AdminPermission)]
+        [Authorize(Roles = Roles.AdminOperaterPermission)]
         public async Task<IActionResult> GetRoles()
         {
             try
             {
                 var id = tokenService.GetClaim(HttpContext,JwtClaims.Id);
-                logger.LogInformation(id);
+                //logger.LogInformation(id);
                 return Ok(await userService.GetAllRoles());
             }
             catch (Exception e)
@@ -145,7 +151,7 @@ namespace Server.Controllers
         [ProducesResponseType(typeof(MessageResponseDTO), StatusCodes.Status500InternalServerError)]
 
         [HttpPost]
-        [Authorize(Roles =Roles.AdminPermission)]
+        [Authorize(Roles =Roles.AdminOperaterPermission)]
         public async Task<IActionResult> CreateUser([FromBody] UserCreateDTO requestBody)
         {
             try
@@ -163,6 +169,8 @@ namespace Server.Controllers
                     Longitude=requestBody.Longitude,
                     SettlementId=requestBody.SettlementId
                 };
+                if (User.IsInRole(Roles.Operater))
+                    user.RoleId = Roles.ProsumerId;
 
                 _sqliteDb.Users.Add(user);
                 try
@@ -202,7 +210,9 @@ namespace Server.Controllers
                 var user = await userService.GetUserById(id);
                 if (user == null)
                     return NotFound(new { message = "User doesn't exists" });
-                if((requestBody.RoleId==Roles.AdminId && user.RoleId==Roles.DispatcherId) || (requestBody.RoleId == Roles.DispatcherId&& user.RoleId == Roles.AdminId))
+                if (user.RoleId == Roles.SuperadminId)
+                    return Forbid();
+                if ((requestBody.RoleId==Roles.AdminId && user.RoleId==Roles.DispatcherId) || (requestBody.RoleId == Roles.DispatcherId&& user.RoleId == Roles.AdminId))
                     user.RoleId = requestBody.RoleId;
                 user.Blocked = requestBody.Blocked;
                 user.Email = requestBody.Email;
@@ -263,6 +273,8 @@ namespace Server.Controllers
             var user = await _sqliteDb.Users.FirstOrDefaultAsync(user=>user.Id==id);
             if (user==null)
                 return NotFound(new { message="User doesn't exists" });
+            if (user.RoleId == Roles.SuperadminId)
+                return Forbid();
             user.Blocked = requestBody.Status;
             await _sqliteDb.SaveChangesAsync();
             return Ok(new {message="User is blocked successfully"});
@@ -284,6 +296,8 @@ namespace Server.Controllers
             var user=await userService.GetUserById(id);
             if (user != null)
             {
+                if (user.RoleId == Roles.SuperadminId)
+                    return Forbid();
                 _sqliteDb.Remove(user);
                 await _sqliteDb.SaveChangesAsync();
                 return Ok(new {message="Deleted user"});
